requirements.txt
anthropic>=0.34.0
tweepy>=4.14.0
py-clob-client>=0.17.0
python-dotenv>=1.0.0
requests>=2.31.0
web3>=6.0.0
eth-account>=0.10.0
